# `StandardScalerTransform` optimization