--- a/etna/transforms/math/sklearn.py
+++ b/etna/transforms/math/sklearn.py
@@ -3,6 +3,7 @@ from copy import deepcopy
 from typing import Dict
 from typing import List
 from typing import Optional
+from typing import Sequence
 from typing import Union
 from typing import cast
 
@@ -226,7 +227,8 @@ class SklearnTransform(ReversibleTransform):
         return df
 
     def _preprocess_macro(self, df: pd.DataFrame) -> np.ndarray:
-        x = TSDataset.to_flatten(df, features=self.in_column)[self.in_column].values
+        in_column = cast(Sequence[str], self.in_column)
+        x = TSDataset.to_flatten(df, features=in_column)[self.in_column].values
         return x
 
     def _postprocess_macro(self, df: pd.DataFrame, transformed: np.ndarray) -> np.ndarray: