## Before submitting (must do checklist)

- [ ] Did you read the [contribution guide](https://github.com/etna-team/etna/blob/master/CONTRIBUTING.md)?
- [ ] Did you update the docs? We use Numpy format for all the methods and classes.
- [ ] Did you write any new necessary tests?
- [ ] Did you update the [CHANGELOG](https://github.com/etna-team/etna/blob/master/CHANGELOG.md)?
<!-- For CHANGELOG separate each item in unreleased section by blank line to reduce collisions -->


## Flame graph(before)
- `mode="macro", inplace=True`
`fit_transform`: 11.66s
![image](https://github.com/etna-team/etna/assets/128227577/403cd777-c163-4f2d-8ac5-c76d3b2598e4)
- `mode="macro", inplace=False`
`fit_transform`: 8.69s
![image](https://github.com/etna-team/etna/assets/128227577/c4e4f561-2021-4e03-8602-ee557652aa37)
- `mode="per-segment", inplace=True`
`fit_transform`: 6.19s
![image](https://github.com/etna-team/etna/assets/128227577/cdc57280-fd47-4838-8d56-d61ca7761cd8)
- `mode="per-segment", inplace=False`
`fit_transform`: 4.74s
![image](https://github.com/etna-team/etna/assets/128227577/da945498-2b12-4012-863a-0a3bb854548b)


## Proposed Changes
<!-- Add a more detailed description of the changes if needed. 
No need for typos and docs improvements  -->
* rewrite `_preprocess_macro` method using `TSDataset.to_flatten`

## Flame graph(after)
- `mode="macro", inplace=True` 
`fit_transform`:  11.66s - > 6.01s (x1.9)
![image](https://github.com/etna-team/etna/assets/128227577/b8cbeff1-65f2-4a8b-999a-c3627c2373d4)
- `mode="macro", inplace=False`
`fit_transform`:  8.69s - > 5.23s (x1.7)
![image](https://github.com/etna-team/etna/assets/128227577/c7b7131e-0ebb-4cfe-a132-363b98950ffb)


## Code to reproduce
```
from etna.datasets import TSDataset, generate_ar_df
from etna.transforms import StandardScalerTransform

scaler = StandardScalerTransform(in_column="target", inplace=True, mode="macro")
# scaler = StandardScalerTransform(in_column="target", inplace=False, mode="macro")
# scaler = StandardScalerTransform(in_column="target", inplace=True)
# scaler = StandardScalerTransform(in_column="target", inplace=False)

df = generate_ar_df(periods=200, n_segments=10000, start_time="2000-01-01", sigma=4)
ts = TSDataset.to_dataset(df)
ts = TSDataset(ts, freq="D")

scaler.fit_transform(ts)
```

## Closing issues
<!-- Link Issue you are closing here. 
Put `closes #XXXX` in your comment to auto-close the issue that your PR fixes (if such). -->